Cleaned up final W9 Auth
