+++ frontend/index.html
@@ -3,6 +3,7 @@
 <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+	<link href="/index.css" rel="stylesheet" />
     <title>Vite + React + TS</title>
 </head>
 <body>
