+++ frontend/src/components/SVGImportExample.tsx
@@ -1,0 +1,5 @@
+import randoLogo from "./assets/rando.svg";
+
+export function SVGImportExample() {
+	return (<img src={randoLogo} className="logo react" alt="React logo" />);
+}
