+++ frontend/vite.config.ts
@@ -14,5 +14,5 @@ export default defineConfig({
     build: {
         outDir: './build',
     },
-    base: "./",
+    base: "./"
 })
